Fixed PostgisTrait to handle GeometryCollection correctly
The geometrycollection migration method inserts the column as a geometry column instead of a geography column. This causes saves to fail since the insert is trying to save a geography instead of a geometry. 

The GeometryCollection class does extend the Geometry class which implements the GeometryInterface that is checked in the if statement to skip model attributes that are not geometries.