Added support for custom float precision to Point.
Constructor reads value from config/postgis.php:precision or defaults to 6 (the default precision for sprintf). It is not anticipated that users will want to customise precision for an individual point, so no argument has been added to the constructor signature.

A setter is provided for precision, though this is primarily to enable tests. If users do need to set precision for an individual point, the setter may be used.

stringifyFloat has been converted to dynamic function to ease testing. It was only ever called from a non-static context anyway.